Big update of semantics to support internal calls
The steps are a lot smaller now, with explicit continuations
representing all the contexts a call might return to.

An alternative approach (which maybe we will want to prove equivalent at
some point?) would be to make statement and expression semantics
mutually recursive, leaving the continuations implicit.

The test theorems are not updated yet (except for one). Execution by
rewriting is a lot slower now so possibly we want to cv_translate the
semantics for execution.